+++ src/main/java/edu/kpi/testcourse/auth/AuthorizationMockService.java
@@ -1,6 +1,6 @@
 package edu.kpi.testcourse.auth;
 
-import edu.kpi.testcourse.exception.user.UnauthorizedException;
+import edu.kpi.testcourse.exception.auth.UnauthorizedException;
 
 /**
  * Mock service to provide user authorization by completely random JWT token.
