+++ src/main/java/edu/kpi/testcourse/exception/auth/UnauthorizedException.java
@@ -1,4 +1,4 @@
-package edu.kpi.testcourse.exception.user;
+package edu.kpi.testcourse.exception.auth;
 
 /**
  * Throw this exception if a JWT token is invalid or not provided.
